Optionally prevent output to stdout/stderr.
Fixes #16.

In general it is not necessary that a test writes to stdout/stderr.
Avoiding this output may speed up the test and reduce the clutter on
the commandline. Your test does not write to stdout/stderr if you
create the rules with the LOG_ONLY log mode.

    new StandardErrorStreamLog(LOG_ONLY);

or

    new StandardOutputStreamLog(LOG_ONLY);

LOG_ONLY has not been choosen to be the default mode in order to keep
the current behaviour if the rules are created with the default
constructor.